서비스 예외처리 로직 개선,Salary 관련 api 생성

+++ src/main/java/com/example/jobproject/service/ApplicationsService.java
@@ -5,6 +5,8 @@ import com.example.jobproject.entity.ApplyList;
 import com.example.jobproject.entity.Recruit;
 import com.example.jobproject.entity.Resume;
 import com.example.jobproject.entity.User;
+import com.example.jobproject.exception.DataNotFoundException;
+import com.example.jobproject.exception.DeadlinePassedException;
 import com.example.jobproject.repository.ApplyListRepository;
 import com.example.jobproject.repository.RecruitRepository;
 import com.example.jobproject.repository.ResumeRepository;
@@ -37,10 +39,10 @@ public class ApplicationsService {
         String username = authentication.getName(); // API 요청에 들어있는 토큰의 username 정보 (JWT방식이라 토큰으로 로그인유무 판단)
         // username을 통해 User 엔티티 조회
         User user = userRepository.findByUsername(username)
-                .orElseThrow(() -> new IllegalArgumentException("해당 유저를 찾을 수 없습니다."));
+                .orElseThrow(() -> new DataNotFoundException("해당 유저를 찾을 수 없습니다."));
         // 이력서 엔티티 조회
         Resume resume = resumeRepository.findById(resumeId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 이력서를 찾을 수 없습니다."));
+                .orElseThrow(() -> new DataNotFoundException("해당 이력서를 찾을 수 없습니다."));
         // 가져온 이력서가 현재 로그인 사용자가 소유한 이력서인지 조회
         boolean isOwnedByUser = false;
         for (Resume r : user.getResumes()) {
@@ -55,14 +57,20 @@ public class ApplicationsService {
 
         // 채용 공고 엔티티 조회
         Recruit recruit = recruitRepository.findById(recruitId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 채용 공고를 찾을 수 없습니다."));
+                .orElseThrow(() -> new DataNotFoundException("해당 채용 공고를 찾을 수 없습니다."));
 
-        // 유저가 해당 채용공고에 이미 조회했는지 확인
+
+        // 유저가 해당 채용공고에 이미 지원했는지 확인
         List<ApplyList> userapplyLists = user.getApplyList();
         for(ApplyList applyList : userapplyLists){
             if (applyList.getId() == recruitId) {
-                throw new DataDuplicateException("현재 로그인한 사용자는 해당 이력서를 소유하고 있지 않습니다.");
+                throw new DataNotFoundException("현재 로그인한 사용자는 이미 해당 공고에 지원했습니다.");
             }
+        }
+
+        //마감된 공고는 취소(삭제) 불가
+        if (LocalDateTime.now().isAfter(recruit.getRecruitDeadline())) {
+            throw new DeadlinePassedException("마감된 공고, 지원 가능한 기간이 만료되었습니다.");
         }
 
         // ApplyList 객체 생성 및 데이터 설정
@@ -78,6 +86,15 @@ public class ApplicationsService {
 
     //삭제
     public void delete(int applyId){
+        ApplyList applyList = applyListRepository.findById(applyId)
+                .orElseThrow(() -> new DataNotFoundException("해당 채용 공고를 찾을 수 없습니다."));
+        Recruit recruit = recruitRepository.findById(applyList.getRecruit().getId())
+                .orElseThrow(() -> new DataNotFoundException("해당 채용 공고를 찾을 수 없습니다."));
+
+        //마감된 공고는 취소(삭제) 불가
+        if (LocalDateTime.now().isAfter(recruit.getRecruitDeadline())) {
+            throw new DeadlinePassedException("마감된 공고, 취소 가능한 기간이 만료되었습니다.");
+        }
         applyListRepository.deleteById(applyId);
     }
 
@@ -90,14 +107,13 @@ public class ApplicationsService {
         String username = authentication.getName(); // API 요청에 들어있는 토큰의 username 정보 (JWT방식이라 토큰으로 로그인유무 판단)
         // username을 통해 User 엔티티 조회
         User user = userRepository.findByUsername(username)
-                .orElseThrow(() -> new IllegalArgumentException("해당 유저를 찾을 수 없습니다."));
+                .orElseThrow(() -> new DataNotFoundException("해당 유저를 찾을 수 없습니다."));
         //유저가 지원한 채용공고 리스트
         List<ApplyList> userapplyLists = user.getApplyList();
         for(ApplyList applyList : userapplyLists){
             String title=applyList.getRecruit().getRecruitTitle();
             list.add(title);
         }
-
         return list;
     }
 }
